Added media set controls to sub_databaseBackup
Added the ability to find a backup file from a media set name and use
that file to append to the current media set.

Modified the audit sp and backup_history_file table to add media set
parameters. The audit sp will now record if the backup is from a new
media set or existing one.